Transform the WIOD data.table into a flat/long data.table

Transform the WIOD data.table into a flat/long data.table

flatWIOD(wiot, rows = seq_len(dim(wiot)[1]),
  columns = seq_len(dim(wiot)[2]))

Arguments

wiot
WIOD data.table returned by loadWIOD

rows
wiot's rows to be dropped (a numeric vector with negative numbers) or kept (a numeric vector with positive) prior to flattening/reshaping; default: all rows kept

columns
wiot's columns to be dropped (a numeric vector with negative numbers) or kept (a numeric vector with positive) prior to flattening/reshaping; default: all columns kept

Value

A data.table where each row corresponds to a combination of a source (exporting) and destination (importing) sectors and countries, with ten columns:

ExpSectorNr

Exporting sector -- WIOD numeric codes for industries (integer vector), see also sectors

ExpCountry

Exporting country -- a three-letter ISO country code, as in the WIOD file (character vector), see also countries

value

An absolute value of the flow (in US dollars as in the WIOD source file)

ImpSectorNr

Importing sector -- WIOD numeric codes for industries and final use sectors (integer vector), see also sectors

ImpCountry

Importing country -- a three-letter ISO country code, as in the WIOD file (character vector), see also countries